requirements.txt
pandas
apache-airflow-providers-apache-kafka